Little corrections
HashDigest: Simultaneous use of int and short
EulerAngles: Use of the macro
Geom: cone was missing

+++ src/Nazara/Core/HashDigest.cpp
@@ -89,7 +89,7 @@ NzString NzHashDigest::ToHex() const
 	return NzString(new NzString::SharedString(1, length, length, hexOutput));
 }
 
-nzUInt8 NzHashDigest::operator[](unsigned short pos) const
+nzUInt8 NzHashDigest::operator[](unsigned int pos) const
 {
 	#if NAZARA_CORE_SAFE
 	if (pos >= m_digestLength)
